feat: split `reth` and `hera` binaries (#23)
Followup from #14 

Many ways to do this, but for now here's what we have:

- the `reth` binary runs the default reth CLI commands with an optional
`--hera` flag. If set, we also try to parse all `HeraArgsExt` flags and
install the ExEx on top of it. The idea is that in the future we may
have more ExExes that can be added in this way.
- `hera` binary is currently empty, but will contain the standalone
rollup node entrypoint.
- refactored `hera-exex` to its own crate, this will make it possible to
also run hera as exex from the `hera` binary as well.

Note: there is a compilation error due to
https://github.com/ethereum-optimism/kona/pull/454, unrelated to this
pr.

Open to feedback