fix(rollup): init pipeline based on real l2 tip (#88)
This PR fixes the pipeline initialization by fetching the L2 chain tip
and its corresponding L1 origin block info.
This also fixes the issue of not initializing the driver `SyncCursor`
which would currently panic otherwise.

The `RollupPipeline` doesn't depend on an `L2ChainProvider` generic
anymore, because we were using the same provider independently of the
context mode. The second reason is because the `latest_block_number()`
method is only available on the `AlloyL2ChainProvider` and we need to
call it to fetch the L2 tip before initializing the pipeline.

If in the future we need to support different L2 providers, I suggest
adding a `L2ChainProviderExt` trait which also contains the ability to
fetch the latest block number, or alternatively upstream this change
into `kona_derive`'s `L2ChainProvider` trait directly.